fix: improve IME input width for MultiSelect component
Fixes #88

This commit resolves the issue where the MultiSelect component's input
field was too narrow when using Chinese input method (IME), causing the
IME candidate box to display incorrectly.

Changes:
- Created MultiSelectEnhanced component that encapsulates IME handling
  logic with per-instance composition state tracking
- Added dynamic width expansion when composing (Chinese/Japanese/Korean
  input), expanding from default to 96% width automatically
- Added smooth transition animation for width changes
- Refactored FilterModal to use the new enhanced component
- Added flex layout support to allow input to grow naturally

The solution detects composition events (compositionstart/compositionend)
and applies a CSS class to dynamically adjust the input width, providing
enough space for IME candidates while preventing layout issues.
